Furkan_Work/ai_detection/beh_unit/requirements.txt
kafka-python==2.0.2
transformers==4.40.0
decord==0.6.0
numpy==1.26.4
huggingface-hub==0.23.1
